reboot/backend/requirements.txt
fastapi
uvicorn[standard]
bcrypt
pyjwt
Pillow
